Abaco 1.6.0

+++ actors/store.py
@@ -6,7 +6,9 @@ import os
 import urllib.parse
 
 import configparser
+import pprint
 import redis
+from pymongo.errors import WriteError, DuplicateKeyError
 from pymongo import MongoClient
 
 from config import Config
@@ -92,193 +94,13 @@ class AbstractTransactionalStore(AbstractStore):
         """Execute a callable, f, within a lock on key `key`."""
         pass
 
-class RedisStore(AbstractStore):
-
-    def __init__(self, host, port, db=0):
-        redis_password = os.environ.get("redis_password", None)
-        if not redis_password:
-            # check in the config file:
-            try:
-                redis_password = Config.get('store', 'redis_password')
-            except configparser.NoOptionError:
-                pass
-
-        self._db = redis.StrictRedis(host=host, port=port, db=db, password=redis_password)
-        try:
-            self.ex = int(Config.get('web', 'log_ex'))
-        except ValueError:
-            self.ex = -1
-
-    def __getitem__(self, key):
-        return _do_get(self._db.get, key)
-
-    def __setitem__(self, key, value):
-        _do_set(self._db.set, key, value)
-
-    def __delitem__(self, key):
-        self._db.delete(key)
-
-    def __iter__(self):
-        return self._db.scan_iter()
-
-    def __len__(self):
-        return self._db.dbsize()
-
-    def set_with_expiry(self, key, obj):
-        """Set `key` to `obj` with automatic expiration of `ex` seconds."""
-        self._db.set(key, obj, ex=self.ex)
-
-    def update(self, key, field, value):
-        "Atomic ``self[key][field] = value``."""
-
-        def _update(pipe):
-            cur = _do_get(pipe.get, key)
-            cur[field] = value
-            pipe.multi()
-            _do_set(pipe.set, key, cur)
-
-        self._db.transaction(_update, key)
-
-    def pop_field(self, key, field):
-        "Atomic pop ``self[key][field]``."""
-
-        # Note: it is problematic to return items from within the _db.transaction() wrapper, as these
-        # We initially had the following implementation, but it did not work: the value returned was a list of
-        # results returned from the redis pipe (e.g. [True]) instead of the actual return value.
-        #     def _pop(pipe):
-        #         cur = _do_get(pipe.get, key)
-        #         value = cur.pop(field)
-        #         pipe.multi()
-        #         _do_set(pipe.set, key, cur)
-        #         return value
-        #
-        #     return self._db.transaction(_pop, key)
-
-        with self._db.pipeline() as pipe:
-            while 1:
-                try:
-                    pipe.watch(key)
-                    cur = _do_get(pipe.get, key)
-                    value = cur.pop(field)
-                    _do_set(pipe.set, key, cur)
-                    pipe.execute()
-                    return value
-                except redis.WatchError:
-                    continue
-
-    def update_subfield(self, key, field1, field2, value):
-        "Atomic ``self[key][field1][field2] = value``."""
-
-        def _update(pipe):
-            cur = _do_get(pipe.get, key)
-            cur[field1][field2] = value
-            pipe.multi()
-            _do_set(pipe.set, key, cur)
-
-        self._db.transaction(_update, key)
-
-    def getset(self, key, value):
-        "Atomically: ``self[key] = value`` and return previous ``self[key]``."
-
-        value = self._db.getset(key, json.dumps(value).encode('utf-8'))
-        if value is not None:
-            return json.loads(value.decode('utf-8'))
-
-    def add_if_empty(self, key, field, value):
-        """
-        Atomic ``self[key][field] = value`` if s``self[key]`` does not exist or is empty.
-        Add a value, `value`, to a field, `field`, under key, `key`, only if the key does not exist
-        or it's value is currently empty. Returns the value if it was added; otherwise, returns None.
-        """
-        def _transaction(pipe):
-            try:
-                cur = _do_get(pipe.get, key)
-                if cur is None or cur == {}:
-                    cur[field] = value
-                    pipe.multi()
-                    _do_set(pipe.set, key, cur)
-                    return value
-                else:
-                    return None
-            except KeyError:
-                # if the key doesn't exist at all, go ahead and set the value.
-                obj = {field: value}
-                pipe.multi()
-                _do_set(pipe.set, key, obj)
-            # the key exists in the store; if it is the value empty, and the field:
-        return self._db.transaction(_transaction, key)
-
-    def add_key_val_if_empty(self, key, value):
-        """
-        Atomic ``self[key] = value`` if ``self[key]`` does not exist or is empty.
-        If the key does exist, returns None.
-        """
-        def _transaction(pipe):
-            try:
-                cur = _do_get(pipe.get, key)
-                # if we are here, the key already exists; return None:
-                return None
-            except KeyError:
-                # the key does not exist, so add the value
-                pipe.multi()
-                _do_set(pipe.set, key, value)
-                return value
-        return self._db.transaction(_transaction, key)
-
-    def pop_fromlist(self, key, idx=-1):
-        """
-        Atomic self[key].pop(idx); assumes the data structure under `key` is a list.
-        
-        :return: 
-        """
-        with self._db.pipeline() as pipe:
-            while 1:
-                try:
-                    pipe.watch(key)
-                    cur = _do_get(pipe.get, key)
-                    value = cur.pop(idx)
-                    pipe.multi()
-                    _do_set(pipe.set, key, cur)
-                    pipe.execute()
-                    return value
-                except redis.WatchError:
-                    continue
-
-    def append_tolist(self, key, obj):
-        """
-        Atomic self[key].append(obj); assumes the data structure under `key` is a list.
-
-        :return: 
-        """
-        with self._db.pipeline() as pipe:
-            while 1:
-                try:
-                    pipe.watch(key)
-                    cur = _do_get(pipe.get, key)
-                    cur.append(obj)
-                    pipe.multi()
-                    _do_set(pipe.set, key, cur)
-                    pipe.execute()
-                    return None
-                except redis.WatchError:
-                    continue
-                finally:
-                    pipe.reset()
-
-    def within_transaction(self, f, key):
-        """Execute a callable, f, within a lock on key `key`. The executable, f, should take a single argument that
-        is the current value under the key """
-        def _transaction(pipe):
-            cur = _do_get(pipe.get, key)
-            f(cur)
-
-        return self._db.transaction(_transaction, key)
-
 
 class MongoStore(AbstractStore):
 
     def __init__(self, host, port, database='abaco', db='0', user=None, password=None):
         """
+        Note: pop_fromlist, append_tolist, and within_transaction were removed from the Redis
+        store functions as they weren't necessary, don't work, or don't work in Mongo.
         Creates an abaco `store` which maps to a single mongo
         collection within some database.
         :param host: the IP address of the mongo server.
@@ -299,17 +121,64 @@ class MongoStore(AbstractStore):
         self._mongo_database = self._mongo_client[database]
         self._db = self._mongo_database[db]
 
-    def __getitem__(self, key):
-        result = self._db.find_one({'_id': key})
-        if not result:
-            raise KeyError()
-        return result[key]
+    def __getitem__(self, fields):
+        """
+        Atomically does either:
+        Gets and returns 'self[key]' or 'self[key][field1][field2][...]' as a dictionary
+        """
+        key, _, subscripts = self._process_inputs(fields)
+        result = self._db.find_one(
+            {'_id': key},
+            projection={'_id': False})
+        if result == None:
+            raise KeyError(f"'_id' of '{key}' not found")
+        try:
+            return eval('result' + subscripts)
+        except KeyError:
+            raise KeyError(f"Subscript of {subscripts} does not exists in document of '_id' {key}")
 
-    def __setitem__(self, key, value):
-        self._db.save({'_id': key, key: value})
+    def __setitem__(self, fields, value):
+        """
+        Atomically does either:
+        Sets 'self[key] = value' or sets 'self[key][field1][field2][...] = value'
+        """
+        key, dots, _ = self._process_inputs(fields)
+        try:
+            if isinstance(fields, str) and isinstance(value, dict):
+                result = self._db.update_one(
+                    filter={'_id': key},
+                    update={'$set': value},
+                    upsert=True)
+            else:
+                result = self._db.update_one(
+                    filter={'_id': key},
+                    update={'$set': {dots: value}},
+                    upsert=True)
+        except WriteError:
+            raise WriteError(
+                "Likely due to trying to set a subfield of a field that does not exists." +
+                "\n Try setting a dict rather than a value. Ex. store['id_key', 'key', 'field'] = {'subfield': 'value'}")
+        if result.raw_result['nModified'] == 0:
+            if not 'upserted' in result.raw_result:
+                logger.debug(f'Field not modified, old value likely the same as new. Key: {key}, Fields: {dots}, Value: {value}')
 
-    def __delitem__(self, key):
-        self._db.delete_one({'_id': key})
+    def __delitem__(self, fields):
+        """
+        Atomically does either:
+        Deletes 'self[key]'
+        Unsets 'self[key][field1][field2][...]'
+        """
+        key, dots, subscripts = self._process_inputs(fields)
+        if not subscripts:
+            result = self._db.delete_one({'_id': key})
+            if result.raw_result['n'] == 0:
+                logger.debug(f"No document with '_id' found. Key:{key}, Fields:{dots}")
+        else:
+            result = self._db.update_one(
+                filter={'_id': key},
+                update={'$unset': {f'{dots}': ''}})
+            if result.raw_result['nModified'] == 0:
+                logger.debug(f"Doc with specified fields not found. Key:{key}, Fields:{dots}")
 
     def __iter__(self):
         for cursor in self._db.find():
@@ -317,37 +186,158 @@ class MongoStore(AbstractStore):
         # return self._db.scan_iter()
 
     def __len__(self):
-        return self._db.count()
+        """
+        Returns the estimated document count of a store to give length
+        We don't use '.count_documents()' as it's O(N) versus O(1) of estimated
+        Length for a document or subdocument comes from len(store['key']['field1'][...]) using dict len()
+        """
+        return self._db.estimated_document_count()
+
+    def __repr__(self):
+        """
+        Returns a pretty string of the entire store with '_id' visible for developer use
+        """
+        return pprint.pformat(list(self._db.find()))
+
+    def _process_inputs(self, fields):
+        """
+        Takes in fields and returns the key corresponding with '_id', dot notation
+        for getting to a specific field in a Mongo query/filter (ex. 'field1.field2.field3.field4')
+        and the subscript notation for returning a specified field from a result dictionary
+        (ex. `['field1']['field2']['field3']['field4']`)
+        """
+        if isinstance(fields, str):
+            key = dots = fields
+            subscripts = ''
+        elif isinstance(fields, list) and len(fields) == 1:
+            key = dots = fields[0]
+            subscripts = ''
+        else:
+            key = fields[0]
+            dots = '.'.join(fields[1:])
+            subscripts = "['" + "']['".join(fields[1:]) + "']"
+        return key, dots, subscripts
 
     def _prepset(self, value):
         if type(value) is bytes:
             return value.decode('utf-8')
         return value
 
-    def set_with_expiry(self, key, obj):
-        """Set `key` to `obj` with automatic expiration of the configured seconds."""
-        self._db.save({'_id': key, 'exp': datetime.utcnow(), key: self._prepset(obj)})
+    def pop_field(self, fields):
+        """
+        Atomically pops 'self[key] = value' or 'self[key][field1][field2][...] = value'
+        """
+        key, dots, subscripts = self._process_inputs(fields)
+        if not subscripts:
+            result = self._db.find_one(
+                {'_id': key},
+                projection={'_id': False})
+            if result == None:
+                raise KeyError(f"'_id' of '{key}' not found")
+            del_result = self._db.delete_one({'_id': key})
+            if del_result.raw_result['n'] == 0:
+                raise KeyError(f"No document deleted")
+            return result
+        else:
+            result = self._db.find_one_and_update(
+                filter={'_id': key},
+                update={'$unset': {dots: ''}})
+            try:
+                return eval('result' + subscripts)
+            except KeyError:
+                raise KeyError(f"Subscript of {subscripts} does not exist in document of '_id' {key}")
 
-    def update(self, key, field, value):
-        "Atomic ``self[key][field] = value``."""
-        result = self._db.find_and_modify(query={'_id': key},
-                                          update={'$set': {'{}.{}'.format(key,field): value}})
-        if not result:
-            raise KeyError()
+    def set_with_expiry(self, fields, value):
+        """
+        Atomically:
+        Sets 'self[key] = value' or 'self[key][field1][field2][...] = value'
+        Creates 'exp' subdocument in document root with current time for use with MongoDB TTL expiration index
+        Note: MongoDB TTL checks every 60 secs to delete files
+        """
+        key, dots, _ = self._process_inputs(fields)
+        if len(fields) == 1 and isinstance(value, dict):
+            result = self._db.update_one(
+                filter={'_id': key},
+                update={'$set': {'exp': datetime.utcnow()},
+                        '$set': value},
+                upsert=True)
+        else:
+            result = self._db.update_one(
+                filter={'_id': key},
+                update={'$set': {'exp': datetime.utcnow(), dots: self._prepset(value)}},
+                upsert=True)
 
-    def pop_field(self, key, field):
-        "Atomic pop ``self[key][field]``."""
-        result = self._db.find_and_modify(query={'_id': key},
-                                          update={'$unset': {'{}.{}'.format(key, field): ''}})
-        result = result.get(key)
-        return result[field]
+    def full_update(self, key, value, upsert=False):
+        result = self._db.update_one(key, value, upsert)
+        return result
 
-    def update_subfield(self, key, field1, field2, value):
-        "Atomic ``self[key][field1][field2] = value``."""
-        self._db.update_one({'_id': key}, {'$set': {'{}.{}.{}'.format(key, field1, field2): value}})
+    def getset(self, fields, value):
+        """
+        Atomically does either:
+        Sets 'self[key] = value' and returns previous 'self[key]'
+        Sets 'self[key][field1][field2][...] = value' and returns previous 'self[key][field1][field2][...]'
+        """
+        key, dots, subscripts = self._process_inputs(fields)
+        result = self._db.find_one_and_update(
+            filter={'_id': key, dots: {'$exists': True}},
+            update={'$set': {dots: value}})
+        if result == None:
+            raise KeyError(f"1Subscript of {subscripts} does not exist in document of '_id' {key}")   
+        try:
+            if len(fields) == 1:
+                return eval(f"result['{key}']")
+            else:
+                return eval('result' + subscripts)
+        except KeyError:
+            raise KeyError(f"Subscript of {subscripts} does not exist in document of '_id' {key}")
 
-    def getset(self, key, value):
-        "Atomically: ``self[key] = value`` and return previous ``self[key]``."
-        value = self._db.find_and_modify(query={'_id': key},
-                                         update={key: value})
-        return value[key]
+    def items(self, filter_inp=None, proj_inp={'_id': False}):
+        " Either returns all with no inputs, or filters when given filters"
+        return list(self._db.find(
+            filter=filter_inp,
+            projection=proj_inp))
+
+    def add_if_empty(self, fields, value):
+        """
+        Atomically:
+        Sets 'self[key] = value' or 'self[key][field1][field2][...] = value'
+        Only if the specified key/field(s) combo does not exist or is empty
+        Returns the value if it was added; otherwise, returns None
+        Note: Will not override a field set to a value in order to create a subfield
+        """
+        key, dots, _ = self._process_inputs(fields)
+        try:
+            if len(fields) == 1 and isinstance(value, dict):
+                result = self._db.update_one(
+                    filter={'_id': key},
+                    update={'$setOnInsert': value},
+                    upsert=True)
+                if result.upserted_id:
+                    return key
+            elif len(fields) == 1:
+                result = self._db.update_one(
+                    filter={'_id': key},
+                    update={'$setOnInsert': {dots: value}},
+                    upsert=True)
+                if result.upserted_id:
+                    return key
+            else:
+                try:
+                    result = self._db.update_one(
+                        filter={'_id': key},
+                        update={'$setOnInsert': {dots: value}},
+                        upsert=True)
+                    if result.upserted_id:
+                        return fields
+                except WriteError:
+                    print("Likely due to trying to set a subfield of a field that is already set to one value")
+                    pass
+            return None
+        except DuplicateKeyError:
+            return None
+    
+    def aggregate(self, pipeline, options = None):
+        return self._db.aggregate(pipeline, options)
+
+    def create_index(self, index_list):
+        return self._db.create_index(index_list)
